Add files via upload
The goal of this project is to develop a predictive model for detecting the pres- ence or absence of heart disease based on various health-related features. It is a binary classification problem where the model aims to accurately classify individuals into two categories: those with heart disease and those without.